refactor: create initialJobsService
